@@ -22,14 +22,12 @@
     "diskusage": "^1.1.3",
     "figlet": "^1.5.0",
     "globby": "11.0.3",
-    "ink": "^3.0.8",
     "inquirer": "^8.0.0",
     "js-yaml": "^4.0.0",
     "line-reader": "^0.4.0",
-    "mustache": "^4.1.0",
+    "mustache": "^4.2.0",
     "network": "^0.5.0",
     "pjson": "^1.0.9",
-    "react": "^17.0.2",
     "shelljs": "^0.8.4",
     "tslib": "^ 2.1.0"
   },
@@ -38,16 +36,15 @@
     "@oclif/test": "^1.2.8",
     "@types/chai": "^4.2.15",
     "@types/clear": "^0.1.1",
-    "@types/figlet": "^1.2.0",
+    "@types/figlet": "^1.5.0",
     "@types/inquirer": "^7.3.1",
     "@types/js-yaml": "^4.0.0",
     "@types/mocha": "^8.2.2",
     "@types/mustache": "^4.1.1",
-    "@types/node": "^14.14.36",
-    "@types/react": "^17.0.3",
+    "@types/node": "^14.14.37",
     "@types/shelljs": "^0.8.8",
     "chai": "^4.3.4",
-    "eslint": "^7.22.0",
+    "eslint": "^7.23.0",
     "eslint-config-oclif": "^3.1.0",
     "eslint-config-oclif-typescript": "^0.2.0",
     "mocha": "^8.3.2",
@@ -99,7 +96,7 @@
     "deb": "oclif-dev pack:deb",
     "postpack": "rm -f oclif.manifest.json",
     "posttest": "eslint . --ext .ts --config .eslintrc",
-    "prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
+    "prepack": "npm shrinkwrap && rm lib -rf && tsc -b && oclif-dev manifest && oclif-dev readme",
     "test": "nyc --extension .ts mocha --forbid-only \"test/**/*.test.ts\"",
     "version": "oclif-dev readme && git add README.md",
     "lint": "eslint . --ext .ts --config .eslintrc",
